change in result.html

+++ my_app.py
@@ -15,7 +15,7 @@ def ValuePredictor(to_predict_list):
   b = ["variable a", "variable b", "variable c"]
   to_predict = pd.DataFrame(a, columns=b)
   for col in b:
-    to_predict[col] = to_predict[col].astype('int')
+    to_predict[col] = to_predict[col].astype('float')
   loaded_model = joblib.load("modelo_2.pkl")
   result = loaded_model.predict_proba(to_predict)[:,1]
   return result[0]
